+++ src/main/java/com/gym/idat/repository/ClaseRepository.java
@@ -3,5 +3,5 @@ package com.gym.idat.repository;
 import com.gym.idat.model.Clase;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClaseRepository extends JpaRepository<Clase, Integer> {
+public interface ClaseRepository extends JpaRepository<Clase, Long> {
 }
