+++ src/main/java/com/gym/idat/repository/PlanpagoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.gym.idat.model.Planpago;
 
-public interface PlanpagoRepository extends JpaRepository<Planpago, Integer> {
+public interface PlanpagoRepository extends JpaRepository<Planpago, Long> {
 
 }
