+++ src/main/java/com/gym/idat/services/ClienteService.java
@@ -31,7 +31,7 @@ public class ClienteService {
 	        return repository.findAll();
 	    }
 
-	    public Cliente getClienteById(int id) {
+	    public Cliente getClienteById(Long id) {
 	        return repository.findById(id).orElse(null);
 	    }
